improve get_api_key() utility method
Instead of requesting the same username and password every time,
the method now checks for a `MUCKROCK_TOKEN` environment variable.
If it cannot find one, it then prompts the user for their username
and their password. Upon obtaining the token, it tells the user
how to add the token to their environment to skip future logins.